requirements.txt
python-telegram-bot[job-queue]==21.6
psycopg2-binary
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.2
websocket-client==1.8.0
beautifulsoup4==4.12.3
lxml==5.1.0
httpx>=0.24.0
google-genai>=0.8.0
